UploadPackTest: refactor capability config test
UploadPackTest.java contains tests that check behavior when
"allowfilter" and "allowrefinwant" are not set, are set, and are not set
but the client insists on using them anyway. Because another capability
is to be included in a subsequent patch, refactor the common code in
these tests.

Remove setBoolean calls with "false", as they are no-ops.

Also take the opportunity to eliminate the overspecification of the
"fetch=" line returned by the capability advertisement.

Change-Id: I289bbd11c902a513cd8d53bc34767e61ebbd5f17
Signed-off-by: Jonathan Tan <jonathantanmy@google.com>